[Refactor] RunningVC로부터 RunningStatusView 분리
